requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
networkx>=3.0
pyyaml>=6.0
python-hcl2>=4.3.0
boto3>=1.26.0
numpy>=1.24.0